requirements.txt
numpy==1.18.5
streamlit_theme==0.58.0
plotly==4.14.3
pandas==1.2.3
streamlit==0.78.0
requests==2.25.1
matplotlib==3.3.4
tensorflow==2.3.0
Pillow==8.1.2
statsmodels==0.12.2